requirements.txt
flask
sqlalchemy
requests
PyYAML
wheel
setuptools
nose
coverage
